Override ReceiveCard method with helper functions
When receiving a card, a BlackjackPlayer must decide if its hand is
above, below, or equal to the targetValue (=21 by default). Hence, we
must override the ReceiveCard() method inherited from Player.

BestPossibleHand() could have been implemented much more simply if
the targetValue was always 21 as it is in Casinos, but I intend to add
in the option to change the targetValue and number of decks later on,
so this method had to be designed to work for any number of decks and
for any unsigned integer targetValue.